Fixed comparison of signed/unsigned in fl2_compress.c

+++ C/fast-lzma2/fl2_compress.c
@@ -581,7 +581,7 @@ FL2LIB_API size_t FL2LIB_CALL FL2_compressCCtx(FL2_CCtx* cctx,
     if (cctx->params.doXXH && !cctx->params.omitProp) {
         XXH32_canonical_t canonical;
         DEBUGLOG(5, "Writing hash");
-        if(end - dstBuf < XXHASH_SIZEOF)
+        if(end - dstBuf < (long)XXHASH_SIZEOF)
             return FL2_ERROR(dstSize_tooSmall);
         XXH32_canonicalFromHash(&canonical, XXH32(src, srcSize, 0));
         memcpy(dstBuf, &canonical, XXHASH_SIZEOF);
